functions/update_priority/requirements.txt
numpy==2.2.1
scipy==1.15.1
networkx==3.4.2
flask==3.1.0
google-cloud
google-cloud-firestore
flask
pandas
functions-framework